<1st Movement and Skelatal Animation>
Implemented the first movement code for Minions, left clicking a minion selects it, and left clicking a space will cause it to move to that directly to that space. Currently this code ignores any spaces in between.

Added a skelatal animation rig for the Minion and a test animation of attacking with a 2 handed melee weapon.