<Implemented Enemy AI>
Implemented Enemy AI to allow them to move to and attack the necromancer and the attack minions that block their path.

Also added placeholder Necromancer, the minion variation that enemies will focus on. Currently its just a normal minion with a purple tinted sprite.

Also commented minion code to make it clearer to understand.