<Added Minion FlipFacing and Dark Energy Meter>
Added FlipFacing function to Minions. By pressing the F key, the player can make the currently selected minion change direction from facing right to facing left and vice versa. This allows minions to target enemies coming from the left side of the screen.

Also added Dark Energy Meter UI. This meter is conposed of a image and sprite object on the UI Canvas, and a seperate Text object on the UITextCanvas(necessary as the sprite masking was interfering with  the Text's visiblity). This meter tracks the players current total of Dark Energy and fills and empties as the amount reachs a set value, this value however is purely visual and there is currently no upper limit on how much dark energy a player can have.

Also switched the necromancer from using hp to using the player's Dark Energy as hp, reducing it whenever he takes damage and only dying should the dark energy drop below 1.