<Added Many UI Elements>
Added Health Meter to display the current and max health of the currently selected minion. Is hidden when the necromancer is selected or nothing is selected.

Added Name Plate to display the name of the currently selected minion. Is hidden when the necromancer is selected or nothing is selected. Also added names to minions.

Added a flip minion facing button to the UI, is hidden when nothing is selected. Clicking this button will flip the facing of the currently selected minion or necromancer from right to left and visa versa.

Added background image to the UI command bar.

Added switch weapons button to the command bar. Is hidden when the necromancer is selected or nothing is selected. Currently has no functionality, that will be added in the next update.

Moved many scripts from assets folder to assets/scripts folder.