<Implemented Mouse Raycasting Override>
Implemented a system to replace Unity's built-in OnMouseDown system with a custom one based out of User_Input_Script. This was necessary as the built in system did not allow for setting layer masks other ignore raycasts, this ment that anything that relied on raycasts for AI logic would block the mouse from clicking on thing behind them. E.g. enemies would block attempts to click the space under them, but the only way to prevent this would make them invisible to raycasts.

This new system runs a raycast at the mouse position whenever the user hits the left mouse button. It assigns this raycast a layer mask based on which mousecommand mode the player is in (e.g. minion select, ability casting, etc.) and uses the layer mask as defined in the inspector of User_Input_Script. The script then calls OnMouseDownOverride on a script on the hit object that implements the MouseDownOverrider Interface.