<Added AbilityTypes, Buff and InstantCast Abilties>
Added AbilityTypes to abilties, which determine the  method by which the ability is cast. Previously the only type was aimCast, whereby the player clicked the button and then clicked a square to cast the ability.

Added InstantCast abilities, these abilities are cast immediately on hitting the ablitity button.

Added Buffs to minions, this system allows for the creation of buffs in the Level_Script_Container.Buff_Database_Script script, these buffs are sets of modifiers that are applied to minion stats. When a minion has a buff applied to their buffs[] list, they will apply any modifiers from the Buff to thier stats AFTER loading their weapon stats. How the modifiers are applied mathmatically is determined by the Buff stat's Calculation enum, e.g. Buff.attackCooldown would be added onto the minion's attackCooldown if Buff.attackCooldownCalculation was set to Add.

Currently Buffs only support the meleeWeaponDamage and attackCooldown stats.

Also added the SprayAndPray Buff and SprayAndPray Instant Cast Ability to apply said buff. This ability applies a 3 second buff to the caster that  set's their attack cooldown to 0.1 seconds.

TODO: replace the series of ability variables with an abilities class that bundles appropriate ability variables together.