<Rebuilt the structure of Weapon, Ability and Buff Database>
Restructured the Weapon, Ability and Buff Database structures to make them all use the same structure. They now have each entry in the database defined as its own field in the script, making it easier to review entries in the database. This change has added a bit more overhead in the scripts when adding new entries to the database but there are now instructures on how to do so in each database, and I believe that the greater clarity when editing entries in the inspector is worth it.

Also reworked projectiles and projectiles buffs so that projectiles will now take their stats(speed, damage, etc.) off of fields defined in the weapon firing them rather than the projectile prefab. Like projectile buffs have been removed from the projectile script and now are applied the weapon stats the projectile is pulling from, just like normal weapon buffs. This change should make it easier to tweak projectile stats by putting them with the rest of the attack stats in weapons. Note that now projectile prefabs only affect the visual or travel behavour of the projectile rather than the statistics.

Streamlined the buff applying code to be more modular.
Fixed buff expiration code to stop removing expired buffs from the list of activeBuffs while iterating through the activeBuffs.