<Added Passive Abilities, Passive Ability UI indicators, UI Tooltips and the open Roster Button>
Added Passive Abilities. This type of ability applies a permanent buff to the minion who uses it.

Added Fleet of Foot, passive ability. This ability passively doubles the movement speed of the minion.

Added Movement speed support to the buffs system.

Also added BaseMovementSpeed as a value to Minion_Movement_Script(Note: this script really should be renamed to Minion_Logic at some point). This will stop speed buffs permantly increasing speed every time they are applied.

Added a Passive Ability UI indicator. This displays a border around ability button to indicate that the button's correspond to a passive ability.

Added Tooltips to the UI. These tooltips display after hovering the mouse over a UI element for aprox 1.5 seconds, displaying the tooltip associated with the element.

Added an Open Minion Roster Button to the UI, currently this button has no function but will be used in the next update to open the roster of minions to summon.

Also applied various minor fixes. Primarily to ability aiming not be overriden when clicking other button. e.g. you would still be aiming a molotov even after clicking on the SprayAndPray ability.