Added Ammo Meter and re-Static-ified Dark Energy Meter
Added Ammo Meter UI. This meter displays the player's current amount of ammo that will be spent to cast certain abilties. Currently no the spending of ammo has not been implemented.

Made dark energy meter methods that were previously not static, static. This was done by implmenting an instance variable, and making all static methods effect that instance of the dark energy meter script. This prevents the need to fetch and store the Dark Energy Meter object in any scripts that use it.