<Static-ified Weapon and Ability Database and Add Ability Icons>
Made various methods from Weapon_Database_Script and Ability_Database_Script static using the instance method(see previous commit), so that scripts relying on them no longer have to  fetch and store a reference to an instance of these scripts.

Also added Ability Icon support. Now ability button's have an image inside of them that is set to the icon of whichever ability is assigned to the button. An abilities icon is defined in its Ability Database entry.