<Added Projectile Scatter Mechanices>
Added support for weapons with scatter projectiles.
By assigned values other zero to a weapon's range array, the weapon can be set to scatter/spread shots up and down spaces.

E.g. A Weapon range array of { 0, 0, 1, 2}  would create a weapon whose shots would hit 4 spaces in the front, as well 1 space up and down from the 3rd space, and 2 spaces up and down from the 4th space.

Added new code in Minion_AI_Script to make them check for enemies that could be hit by their weapon's scatter, when checking wheter or not to fire.

Added new code to Projectile_Logic_Script to make the projectiles split when appropriate, based on the weapon Range inherited from the weapon when they are fired.

Tweaked some code in Space_Script.findGridSpace(). Added an extra optional parameter that allows calls to the method to request it return an approriate end space if a regular grid space cannot be found matching the passed co-ordinates.