<Implemented Y-Based Sprite SortingLayer ordering>
Implemented code to alter the order in sortingLayer of minion and enemy sprites, so that  minions/enemies higher on the grid will be rendered behind minions/enemies lower on the grid.

Also removed an obsolete AbilityDatabase field in Enemy_AI_Script.

Also added a todo list text document detailing what still needs to be done in the project.