<Added Map Screen>
Added new scene to handle the game's map screen where all out of combat actions will be taken.

Added a background image for the map.
Added a border for the background image of the map
Added edge-scrolling to their map's camera.
Added Map Icons, these prefabs display a clickable picture that when clicked will display the journal interface and load a specific scenario to the journel.

Added the foundations of the journel and scenario system.
This system is the means by which out-of-combat encounters are handled via text and buttons. The journel consists of a book with next-page and previous-page buttons and left-page and right-page text boxes. The pages of the journel take their contents from the currrently loaded scenario.

Scenarios are defined in the Journal Prefab's Scenarios_Database_Screen where they contain a large rich text string. These strings can include markups for page breaks using the format [PAGE BREAK] and markup for adding a button to the text that link to another scenario, the markup for buttons are as such:
[BUTTON](scenario name){Text to display on the button}.