<Implemented Finish Battle Button, Singleton for Level_Script_Container and Horde Json File Save/Load>
Implemented the Singleton pattern for Level_Script_Container and all of the scripts that are attached to it. As Level_Script_Container is supposed to exist in every scene AND persist from scene to scene, this pattern ensures that when switching to a new scene, the Level_Script_Container in that new scene will see that one already exists(having persisted from the previous scene) and self-destroy. This will make debuging easier as we can start debuging from any scene without worring about the Level_Script_Container having conflicts/not existing.

Also implemented battle finishing mechanics and the Finish Battle button. Now when a horde as spawned all of its enemies and all of those enemies are dead. The finish battle button appears allowing the player to return to the map screen.

TODO: Implement Freeing the enemies from conversion ritual if there are no other enemies left.

TODO: Limit conversion ritual to one target at a time. Currently the player can attempt to convert as many targets as they like.

Also added the foundations for a .json file based saving and loading system for Hordes. Currently this system operates by hitting F11 in the battle scene to save the current horde as a json file and hitting F10 to load the file and display debug information about it.

The horde json files are stored in peristantdatapath/databases/hordes, and can be edited externally.

The future intenion of this load/save json system is to allow for the storage of database contents as .json files to allow for the easy addition, remove and editing of individual Hordes, Events, Weapons, etc.