>
Added the first parts of Player Inventory/Minion Load Menus. These menus are used for editing minion equipment, names and cosmetics in the player's roster while on the map screen. So far the minion roster part of the inventory has been implemented and loading a minion into the minion viewer.

The minion viewer displays info about a minion selected from the player's roster. It displays a preview of the minion, their name, their abilities, weapons and cosmetics. It includues buttons that will(in a future update) be used for changing what weapons they have equiped, and what cosmetics they have equiped.

Also added a name field to cosmetic and weapon classes, to store a human-readable name for use in menus.

Moved list of Minions and MinionEntry from  Minion_Roster_Script    to Player_Inventory_Script.

Also added code to automatially load all enemy hordes from all the JSON files stored in the persistant storage folder databases/hordes.

Moved the debug command dummyPopulateRoster with premade minions to Inventory_UI_Script and made it trigger when the user hits the insert key instead of on loading the battlescene.