<Implemented Player Map Marker and moving between nodes>
Implmented a Player map marker to display the player's current location on the map. Also added the player having a location on the map. Also implmenting the player moving between connected nodes to trigger scenarios instead of just clicking on any node.

Added code to make the camera in the map screen lock to and follow the player marker whilst it moves between nodes.