<Added Map Persistance>
Fixed a bug where the player's progress on the map would reset to the start when returning from battle scenes. This was caused by the map scene having no persistance mechanics.

Added Map_State_Storage_Script to Level Script Container object. This script is responible for saving and loading the state of the map scene when switching to and from battle scenes.

Modified Map_Icon_Script to include a NodeID string field to aid in storing data on the Map Icon.

Modified Map_Icon_Script  to switch ConnectedNodes from a list of gameObjects to a list of MapIcon ID strings.

Added Player_Map_Marker.setCurrentMapNode() to allow the setting of the players position on the map without triggering the normal movement mechanics(i.e. moveToNewMapNode())

TODO: Reset map screen to default state on gameover/new game.