<Map Screen Bugfixes>
Fixed a bug where the map screen's progress was persisting between new games. Now hitting new game does not load the previous state of the map and calls GenerateMap(Currently an empty method that does nothing).

Also made it so that only map nodes the player could possibly move to will highlight on mouseOver.

Added [EXIT BUTTON] markup tag to scenarios, this markup generates a button to close the current journal entry and return to the map.

KNOWN BUG: generateExitButton() causes out of bounds errors with the button positioning method when trying to place the new Exit Buttons in certain positions(e.g. at the end of the text body), resulting in the button failing to be properly placed in the text.