<Build Pre-Alpha Map>
Added code to generate a fixed Map for the prealpha.

Added scenarios  and hordes to support the preAlpha map. Stored backup copies of these scenarios in the .json backups folder.

Reworked Exit Button placment code.

Reworked Journal Button prefab to scale horiziontally to fill the space on the page, and scale vertical with the lenght of the text.