<Pre-Alpha Prep 1>
Went over the project fixing problems in preparation for Pre-Alpha release.

Added Necromancer1 projectile, weapon and animations. This is the weapon that the necromancer uses. Note this weapon's projectile has it's animator and sprite renderer in a child object instead of the main object, so as to position the sprite offset from the actual collesion box. This change will likely be applied to other projectiles in later commits.

Reworked Projectile code to work if the sprite and animator are placed as a child object of the projectile object.

Fixed a bug where inventory weapon buttons were being generated without images, caused by the inventory buttons pulling their image from the weapon's weapon sprite instead of it's icon sprite.

Completed Pre-Alpha player profile, so that it now contains a proper inventory of weapons and a proper pre-populated roster.

Added a 5 coloumn of gridspaces to the battlescene grid.

Fixed a bug with some gridspaces and minion AI. Was caused by the presence of 2 copies of the gridspace stacked ontop of each other with indentical script values.