<Pre-Alpha Prep 3>
Increased the size of the Name Input field in the minion loadout.

Corrected numerous spelling, grammer and formating mistakes throughout the pre-alpha scenarios' text.

Fixed a bug where minions could be summoned on the grid space occupied by the necromancer.

Fixed a bug where tooltips would extend outside the tooltip boxes.

Fixed a bug where journal buttons had insufficent padding on the left side, resulting in text blending with the edge of the button.

Fixed a bug where players could click on map icons behind the journal when it was displaying the post-battle screen.

Fixed a bug where the pre-battle screen would not show all the enemy types in the coming battle, if enemy types shared part of their name with another enemy type.