<UI Rework 4>
Finished reworking  UI to work with multiple screen sizes.

Identified the major problem with the ui test. The testing machine was using an extra wide aspect ratio, the builds will now not default to native resolution for the pre-alpha.

Rewrote several scenario json files to work better with the new ui scaling.

Fixed a bug on the map screen where link lines were being drawn in front of map icons.